pyproject.toml
[tool.poetry]
name = "eating-select-bot"
version = "0.0.1"
description = "A eating bot in Discord"
authors = ["Nesquate"]
readme = "README.md"
packages = [{include = "eating_select_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.1.0"
python-dotenv = "^0.21.0"
selenium = "^4.6.1"
beautifulsoup4 = "^4.11.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
